+++ Woof/Woof/Model/Sitter/Sitter+Decoder.swift
@@ -12,5 +12,6 @@ extension Sitter {
         bio = try container.decode(String.self, forKey: .bio)
         rating = (try? container.decode(Double.self, forKey: .rating)) ?? 0
         pricePerHour = (try? container.decode(Double.self, forKey: .pricePerHour)) ?? 0
+        city = (try? container.decode(String.self, forKey: .city)) ?? ""
     }
 }